+++ bot/src/ghutils/core/bot.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from typing import Any
 
 import yaml
-from discord import Color, CustomActivity, Intents, Interaction
+from discord import Color, CustomActivity, Emoji, Intents, Interaction
 from discord.app_commands import AppCommandContext, AppInstallationType
 from discord.ext import commands
 from discord.ext.commands import Bot, Context, NoEntryPointError
@@ -21,7 +21,7 @@ from ghutils.utils.imports import iter_modules
 from .env import GHUtilsEnv
 from .translator import GHUtilsTranslator
 from .tree import GHUtilsCommandTree
-from .types import LoginState
+from .types import CustomEmoji, LoginState
 
 logger = logging.getLogger(__name__)
 
@@ -50,6 +50,7 @@ class GHUtilsBot(Bot):
         self.engine = create_engine(self.env.db_url)
         self.start_time = datetime.now()
         self.language_colors = self._load_language_colors()
+        self._custom_emoji = dict[CustomEmoji, Emoji]()
 
     @classmethod
     def of(cls, interaction: Interaction):
@@ -139,3 +140,34 @@ class GHUtilsBot(Bot):
         gh_default: Color = Color.from_str("#1B1F24")
 
         return self.language_colors.get(language, gh_default)
+
+    async def fetch_custom_emojis(self):
+        logger.info("Fetching custom emojis")
+
+        application_emojis = {
+            emoji.name: emoji for emoji in await self.fetch_application_emojis()
+        }
+
+        self._custom_emoji.clear()
+        for custom_emoji in CustomEmoji:
+            if emoji := application_emojis.get(custom_emoji.name):
+                self._custom_emoji[custom_emoji] = emoji
+            else:
+                logger.warning(
+                    f"Failed to find custom emoji (try running `sync emoji` command): {custom_emoji.name}"
+                )
+
+    async def sync_custom_emojis(self):
+        logger.info("Syncing/uploading custom emojis")
+        self._custom_emoji.clear()
+        for custom_emoji in CustomEmoji:
+            self._custom_emoji[custom_emoji] = await self.create_application_emoji(
+                name=custom_emoji.name,
+                image=custom_emoji.load_image(),
+            )
+
+    def get_custom_emoji(self, custom_emoji: CustomEmoji) -> Emoji:
+        emoji = self._custom_emoji.get(custom_emoji)
+        if emoji is None:
+            raise ValueError(f"Failed to find custom emoji: {custom_emoji.name}")
+        return emoji